all: use host.local.run() instead of run()
host.Host's run() implementation does conceptially the same thing as
common_dpu.run(). It however aims to do the the best way. Whether it
succeeds at that, can be questioned, but the point is that it's an
implementation that is reused and supposed to be as good as possible
(considering various conflicting constrains).

Use it, and don't reimplement it.

The main benefit is that it aims to do useful logging. It's arguably
better than what common_dpu.run() did previously, and it should not
be re-implemented.

Also use shlex.quote() at various places where we paste string arguments
into a shell script.